square, root, exponentiation, and inverse
