+++ src/main/java/com/zipcodewilmington/scientificcalculator/SciFeatures.java
@@ -1,0 +1,6 @@
+package com.zipcodewilmington.scientificcalculator;
+// Liu: Please develop methods of scientific features here
+public class SciFeatures {
+
+
+}
